Create Blgonia
The program reads a matrix of size N, with each cell composed of either "#" or ".". Then, given a coordinate (x,y), the program exchanges the cells directly above, below, or to the side that contain "." to a "x" .  This goes on until all cells that contained an "." and were adjacent to a cell containing "x" are swapped to a cell containing "x". This program simulates the contamination of water sources.